pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
pypi = "^2.1"
schoolopy = "^0.2.2"
Flask = "^2.1.2"
Flask-Cors = "^3.0.10"
python-dateutil = "^2.8.2"
Django = "^4.0.4"
Flask-JSGlue = "^0.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"